create user_login view. this view gets the username abd password from the user using the LoginForm. and then Authenticates the user against the database. snippet rewiew: instantiate the form if the user is submitting the name and password using rquest.POST. check if the form is valid. if it is authenticate the user using the authenticate method using the request object and the name and pasword retireved from the form.
the authenticate method returns None or a user object if the user is authenticated.

login the user:
if user is not none and the user is active, use the login method to login the user, this method also uses the request object and the user object. otherwise display raw httpresponses  or ultimately instantiate a form for user to login.