Add UserEditForm and ProfileEditForm.
UserEditForm: This will allow users to edit their first name, last name,
and email, which are attributes of the built-in Django user model.

ProfileEditForm: This will allow users to edit the profile data that you
save in the custom Profile model. Users will be able to edit their date of
birth and upload a picture for their profile.